requirements.txt
aiohttp>=3.9.0
PyYAML>=6.0
aiosqlite>=0.19.0
